Adds multiple solvers support, defaults to running 3 solvers per solve
The DecisionFormatter is updated and refactored to allow multiple
solvers to run in separate threads for each "solver run". The first
solver to finish is used for the solution if it succeeds, and for the
error if it fails unless ignoring errors is enabled on that solver.

Three solvers are configured and run by default: one as specified on
the command line, one with build key impossible checks enabled, and
one with all the impossible checks enabled. The --disable-multi-solver
command line flag is added to disable this and run only one solver.

Signed-off-by: David Gilligan-Cook <dcook@imageworks.com>