Make RepositoryPackageIterator do stubs in a second pass
Checking embedded stubs in a solve can massively increase the search space
and it is best to avoid trying them unless absolutely necessary. Try all
the versions and builds of non-stubs first, and if none of those work, only
then try all the versions and builds again, but stubs only on this second
pass.

Signed-off-by: J Robert Ray <jrray@imageworks.com>