Fix `generate_binary_build` build digest calculation
Calculate the build digest based on the user-provided options and the
variant being built, but not with any of the resolved options from the
solution (build_env), so that the build digest is deterministic.

There is a gotcha now for tests that `make_build!` needs help knowing what
options are varying so it generates a unique build digest. This only
affected two tests.

Signed-off-by: J Robert Ray <jrray@imageworks.com>