Fix `generate_binary_build` build digest calculation
Calculate the build digest based on the user-provided options and the
variant being built, but not with any of the resolved options from the
solution (build_env), so that the build digest is deterministic.

Add test for build digest calculation

Verify that repeat builds of the same recipe produce a build with the same
digest, even as the versions of its dependencies evolve over time.

Signed-off-by: J Robert Ray <jrray@jrray.org>