Remove runtime-compat-0.33 feature
This backwards compatibility code has been only partially enabled for a
while due to the crate restructuring and it missing from sub-crates. Since
we've not relied on it working for a while and no longer wish to support
running the python version of spk, remove the feature and related code.

Signed-off-by: J Robert Ray <jrray@imageworks.com>