Modest performance improvement for compute_object_manifest
For our large environments with 100+ layers, the layer merging phase takes
noticeable time. As the manifests get larger, cloning them to turn them
into an `Object` gets expensive.

This change delays writing the manifests to storage until they are fully
merged and avoids the expensive clones. In an example case, it shortens the
overall time from 8 seconds to 5 seconds.

Signed-off-by: J Robert Ray <jrray@imageworks.com>