Use `daemon` when launching spfs-monitor
At least at SPI, for not good reasons, when killing a frame on the render
farm, the process tree is walked and processes found are killed with
SIGKILL, giving them no chance to clean up. This would include spfs-monitor
because it retained the pid of spfs-enter as its parent. Using `daemon`,
a double fork makes it detach entirely and become parented to pid 1.

Calling `daemon` still does what `setsid` does, and more. It is assumed
that this change might help with the zombie runtime problem described in
#581.

Signed-off-by: J Robert Ray <jrray@imageworks.com>