Reduces amount of logging from custom logging targets
This removes some of the log messages for the impossible request checks. The default logging level is set to error, as are the custom logging target names. Log levels can still be configured via SPK_LOG.

A basic logging and verbosity documentation page is added to show the
impact of the various settings.

Signed-off-by: David Gilligan-Cook <dcook@imageworks.com>