Fixes bugs assmebling error message when a file is removed during a build.
Changes resolved layers walk processing to not prepend '/spfs' to the
relative paths so the file paths stored in the file and diff mappings
are both relative to /spfs and thus match during lookups. Changes the
single file case to gracefully handle when a filepath is not in the
files_to_packages mapping.

Signed-off-by: David Gilligan-Cook <dcook@imageworks.com>