Run 2 solvers for all solves
The first solver is the one configured on the command line. The second
is based on the first solver but with all the impossible requests
checking enabled.

Solver output is shown for the command line configured solver, and
hidden for the other solver. The '--solver-output-from <name>' option
can be used to select which solver output is shown.

Signed-off-by: David Gilligan-Cook <dcook@imageworks.com>