Detect when pid events are not coming in
Redesign the channel receive loop to use a `Stream` instead, to take
advantage of the utilities that streams provide.

First, filter the stream down to only pid events that pertain to us. Then,
add a timeout to that stream to know when no pid events have shown up for
a while. Also, chunk the events to not spam the log, since these events
can be extremely numerous.

The hash set being used to track pids has changed to a dashmap::HashSet in
order to share the set between the various stream processing functions.

Signed-off-by: J Robert Ray <jrray@imageworks.com>