Update logging configuration for sentry
Rework how logging is configured so the logging can be filtered differently
for different layers, specifically to allow a new target called
"sentry" to exist and be logged at `INFO` level without those log lines
also appearing in the console output.

Use the "sentry" target when it is desired to send an error to sentry but
not have this show up as an "error" message for the user.

Signed-off-by: J Robert Ray <jrray@imageworks.com>