Alternate implementation for payload repair
In this implementation, instead of threading a "pull_from" RepositoryHandle
down the call stack and handling the payload repair inside FSRepository, it
uses a new Repository implementation very similar to Proxy where it has
a primary repo and secondary repos, and will use the secondary repos to
copy missing payloads from if any are detected as missing during
open_payload.

A new trait LocalRepository is needed to make the new Renderer type
generic so it can work on either an FSRepository or the new
PayloadFallback repository wrapper.

Signed-off-by: J Robert Ray <jrray@jrray.org>