Repair missing payloads while rendering
As touched on in #545, if during rendering a payload is discovered to be
missing, and the caller has provided a repo, attempt to pull the missing
payload. This is similar to `spfs check --pull`. The "origin" repo is
provided when possible, but it is not an error if there is no "origin"
repo configured (or enabled).

Signed-off-by: J Robert Ray <jrray@imageworks.com>