Adds repo name and highest versions to solution output
The repo name is added to the front of each resolved package in the
solution, so their format becomes: 'repo/package/version/build'. Whether
a resolved package is the highest versions, or not, is displayed after
the package in a separate column. The highest version info is also
added to 'spk info' output inside a spk runtime environment.

Signed-off-by: David Gilligan-Cook <dcook@imageworks.com>