Add `spfs runtime prune`
One and only strategy currently is to remove runtimes that have a creation
time from before the current boot time. The goal is to have different
strategies that are individually enabled on the command line. With no
flags enabled, nothing happens.

This is a mashup of the `runtime ls` and `runtime rm` commands and there
could be some better code sharing here...

Signed-off-by: J Robert Ray <jrray@imageworks.com>