Fix test that was failing for the wrong reason
I've included a way to log solver output to tracing, and then a tracing
layer that allows the events to be introspected after the fact. This
way, the test can properly detect the failure case that it was looking
for. Additionally, I've had to republish the recipe within that test to
properly create the desired state where a source build is tried and
fails

Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>