Pass --python-version when pip converting packages
Without this, despite --abi, it ends up using the python interpreter being
used (3.7) to resolve compatibility, no matter what --python-version is
passed to spk-convert-pip, making things difficult for those of us still
suffering with python 2.7.

The help for --abi in `pip download --help` mentions --python-version
being needed along with a few other things that default to the current
interpreter. The other ones are probably okay to default that way though.

Signed-off-by: J Robert Ray <jrray@imageworks.com>