Write payload files with custom file permissions
Save up to 50% of disk usage in a local spfs repo by being able to hard
link from a rendered file back to the payload.

Typically a payload file is written with perms 0666. Before the render
proxy file concept was added, this payload file will get chmod'd to
whatever mode the render wanted at the time of rendering, but the problem
was that multiple files in a render (with different permissions) might
share the same payload if they have the same file contents. But a precedent
was established that payload files could have other modes besides 0666.

While rendering, if a payload were to have the same ownership and mode as
what the render needed, then if the render was a hard link back to the
payload file, then a copy of the file isn't required, saving that space.

This commit does two things:

1. when writing a payload to storage, set the mode of the payload to
   whatever mode was attached to the entry that triggered writing the
   payload. This unfortunately requires passing the value down through
   layers of abstraction that otherwise didn't need to know this value.
2. when creating the render proxy file, when conditions are met, hard link
   from the payload file to the proxy file rather than making a copy.

For a single user using an spfs repo, this will effectively cut disk usage
in half. For multiple users sharing an spfs repo, this will reduce the
number of copies of payloads by one.

Signed-off-by: J Robert Ray <jrray@imageworks.com>