Adds traits to Validators interface to separate data the validators access.
This adds two traits for validation data, one for PkgRequests, and one
for Pkg and VarRequests and their associated resolves. The ValidatorT
trait gains an extra method for PkgRequest only validation checking,
and the validators are updated to use the new traits. Three helper
classes are implemented to adapt State, PkgRequest and Test data to
use those traits and allow data to be used in validation.

Signed-off-by: David Gilligan-Cook <dcook@imageworks.com>