Add payload hard link sharing between users (opt-in)
For environments that are going to use spfs in a read-only fashion and not
create spk packages, for example render farm jobs, it is less important
that all the files in /spfs be owned by the current user. In order to save
potentially a huge amount of disk space on a render farm node that is used
by a variety of different users, allow payloads to be shared between users.

This is only safe to enable in some contexts so it is opt-in only and
enabled via a new configuration option in the spfs.conf file. When enabled,
spk disallows building packages to avoid any potential of this behavior
change producing bad packages. With more rigerous testing it may be
possible to prove building packages is actually safe with this enabled, or
fix whatever is making it unsafe.

Signed-off-by: J Robert Ray <jrray@imageworks.com>