Stop treating AlreadyExists from rename as success
For files...

rename replaces a file if it is already exists and returns Ok in that case;
it does not return EEXIST.

I couldn't find a case where rename returns EEXIST (ubuntu 20.04).
Attempting to rename a file onto a empty/non-empty directory fails with
EISDIR.

For directores...

The semantics for renaming directories is different, so add a code comment
where it is possible that the error to rename should be ignored. Add
checking for ENOTEMPTY too because in experimentation (again, on ubuntu)
that's the actual error returned for the situation we're trying to detect.

Signed-off-by: J Robert Ray <jrray@jrray.org>