Adding Impossible request checking to the solver (#506)
Impossible request checks are added in three places: 
- as extra validation before using a build to resolve the current request, 
- to BuildKeyEntry and BuildKey generation to prioritise builds that do no generate an impossible request, 
- and on the initial requests before the first solver step. 
They are disabled by default but can be enabled separately, or all at once, by command line flags and env vars.

The ImpossibleRequestChecker is made thread-safe by using DashMap for its caching and lower-level caching and Atomic counters. The ImpossibleRequestChecker uses tasks, threads, and message channels to reduce the time overhead of the checks. The requirements from the build are still checked sequentially, but each one sets up a channel and launches one task, the maker task, to generate a task per version. The version tasks read and check all the builds for that version, and send a message on the channel as soon as they find a either valid build, or no valid builds. The tasks are all spawned and run in parallel on other threads, and they are stopped as soon as a valid build is found.

There are variations in the way the impossible checks are instigated:
- The initial request checks set up a task per request and runs them concurrently in the current thread, via async. Each makes a single call to the impossible request checker (which will launch more tasks in parallel behind the scenes as described above).
- The build keys checks set up a task per build and runs them in parallel in separate threads. Each makes a single call to the impossible request checker (which will launch more tasks as above).
- The pre-resolve build validation check doesn't make any additional tasks.

Based on current testing, this combination of tasks/threads seemed to perform the best compared with running the impossible checks entirely sequentially. The slowest part of the process is reading the package specifications and the package's list of build component names from disk. The checks speed up dramatically once the all data is cached.

This also includes:
* Adding a basic logging and verbosity documentation page is to describe the impact of the various settings.
* Making pluralize function generic
* Adding traits to Validators interface to separate data the validators access.

Signed-off-by: David Gilligan-Cook <dcook@imageworks.com>