Add progress bar support for spk build and env commands
This involved a refactor where all the rendering logic was moved to the
new Renderer type. In order to support calling through this type the
ManifestViewer trait was removed entirely, requiring instead that an
FSRepository instance is used, as this is the only repo type that can
realistically support renders, anyway. There is some refactoring here,
but overall I think it removes and clarifies some of the code.

Additionally, we allow many layers to be rendered at once with the same
renderer, improving render times for environments with a large number of
small layers or a mix.

Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>