Add fuse filesystem implementation using fuser
The filesystem 'backend' is tracked in the runtime configuration and the
runtime initialization process accounts for this selection when setting
up. The new spfs-fuse command is responsible for setting up the
filesystem, and the spfs-vfs crate defined the virtual fs logic and fuse
implementation via the fuser crate.

The runtime config now also holds a list of secondary repositories that may
be used to read data from, and these repositories are passed along to
the fuse command so that it can leverage the stack and not need to
localize data to a single local repo

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>