Run multiple solvers at once, 2 by default (#568)
The DecisionFormatter is updated and refactored to allow multiple
solvers to run in separate threads for each "solver run". The first
solver to finish is used for the solution if it succeeds, and for the
error if it fails unless ignoring errors is enabled on that solver.

Two solvers are configured and run by default. The first solver is 
the one configured on the command line. The second is based on 
the first solver but with all the impossible requests checking enabled.

Solver output is shown for the command line configured solver, and
hidden for the other solver. The '--solver-output-from <name>' option
can be used to select which solver output is shown.

Signed-off-by: David Gilligan-Cook <dcook@imageworks.com>