Introduce a BlobRead trait with extended attributes
This allows a set of permissions to be attached to the blob data if it's
relevant, while also allowing the use of these permissions to be
optional in the storage implementation without further complicating the
function call.

This change does mean that the check command won't be using entry
permissions. It could be updated to do so like the sync command, but
this felt unnecessary given the situations under which check usually
repairs data.

Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>