Wrap sentry initialization in OnceCell
Prevent the possibility that configure_sentry can be called multiple times
and attempt to initialize sentry more than once.

Add a `shutdown_sentry` function to tear down the value in the OnceCell so
that any cleanup that would happen when the sentry guard is dropped still
has an opportunity to run.

Signed-off-by: J Robert Ray <jrray@imageworks.com>