Report render statistics from spfs-enter
Collect the render summary from the output of spfs-render and report the
summary to sentry. Since this report happens from spfs-enter, extra
context like the spfs-enter command line is available.

This was originally designed when there were multiple calls to spfs-render
for any one spfs-enter so it was necessary to aggregate the summary from
all the spfs-render calls. Now there is a single spfs-render call, but
still there is more context available from the spfs-enter process. It may
be worthwhile to push the statistics even higher for more context.

spfs-enter previously didn't initialize sentry because it is supposed to
be a single-threaded process, but after the runtime has been initialized
then it is safe to go multi-threaded.

Signed-off-by: J Robert Ray <jrray@imageworks.com>