Require has_object implementation
There was one default implementation for has_object, which read the
entire object to determine its existance. This is becoming an issue in
network-based and slow-disk environments, where for larger objects this
seeminly cheap function call can take quite some time. This change
removes the default implementation and requires that repositories
implement this function - as well as adding a cheaper check for
the filesystem implementation.

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>