Require has_payload implementation
Similarly to objects, the default implementation of this required
opening a reader for the payload, which can be an unreasonably heavy
operation in client/server based setups. This change requires a cheap
implementation for all repositories, and updates the existing
has_payload rpc request to have the same structure as has_object.

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>