Improve current_env performance with bulk actions
The slowness of current_env comes from the find_layer_by_filename call
performed for each component of each package in the runtime, since it has a
relatively large overhead each time it is called. To speed up bake, add
a `find_layers_by_filenames` function that performes the same action, but
can find layers for multiple filenames in a single pass over all the
layers.

To make use of this, a new `read_components_bulk` method on the runtime
storage format is used to figure out all the components' "*.cmpt" filenames
that need to be looked up together.

For a large environment with 405 layers in it, this has gone from
originally over 2 minutes, to 17 seconds with the previous commit's
parallelism, to 0.67 seconds with this commit's change.

Signed-off-by: J Robert Ray <jrray@imageworks.com>