Add traits around object digest calculation
To make it possible to inject different digest calculation implementations
into the different types of object, move the `digest` method out of
`Encodable` into its own new trait called `Digestible`.

Add a generic type parameter to each of the objects to hold the digest
calculation to use for the object, supplying a default.

Implement `Digestible` for each object using the generic type.

Signed-off-by: J Robert Ray <jrray@jrray.org>