Create a new Platform object kind
This changes the existing Platform object to be considered `PlatformV1`,
and introduces a new `PlatformV2`. The original Platform uses the same
digest calculation strategy as before, and the new one uses a strategy
that includes its "kind" so there won't be digest collisions between
different kinds of objects.

In order to retain backwards compatibility with existing content in an
spfs repo, the "V1" objects can still be read and synced to another repo
and will retain the same digest they had, since they use the same digest
calculation. Any newly created platform objects will be created as "V2"
objects with the new digest calculation strategy.

The same treatment is needed for `Manifest`, `Layer`, and `Tree`.

Signed-off-by: J Robert Ray <jrray@jrray.org>