Return RenderResult from resolve_and_render_overlay_dirs
Another case where this type is more suitable than a tuple, however in
this case the return value of render_via_subcommand may not reliably return
a valid RenderResult, e.g., if there is some version mismatch between
spfs binaries. That function now returns Option<RenderResult> so it can
handle this situation. This possibility is unexpected, but not being able
to parse the json blob returned from spfs-render doesn't have to be a hard
error.

Signed-off-by: J Robert Ray <jrray@jrray.org>