Adds sync_time and render_time support to spfs enter commands.
The render_time is calculated inside spfs enter for the total time
spend rendering layers to the spfs runtime. The sync_time is passed to
spfs enter via the --synctime command line arg, e.g. by a spfs run
command that has synced some layers locally.

Under the sentry feature, both times are added to the data sent to
sentry.  This also adds the --metrics-in-env option to put the render
summary data and times into a json blob in an environment variable for
other processes to access and use, e.g. send to an external metrics system.

Signed-off-by: David Gilligan-Cook <dcook@imageworks.com>