Updates 'spk info <package>' to run without solving for the package
Updates the info command to look up the package, version, build spec
directly from the given package request, without running a full solve.

For backwards compatibility this adds the '--full-solve' option to
allow the old solving method to be used. But the expectation is that
this will be removed at some point in the future.

Signed-off-by: David Gilligan-Cook <dcook@imageworks.com>