Don't use expect when launching tcsh/csh
Although it was mostly interactive, the use of expect added noise to the
output and muted some shell features. For example, certain key codes
like reverse searching would not work, history was not preserved as
expected, and other inputs may fail to propagate.

This change removes the use of expect in favor of another trick where we
name the csh startup script to the special .cshrc and then set HOME to
the spfs runtime dir where that script is saved. Upon loading, this
script immediately restores the HOME variable and continues as usual.

This trick has been used successfully and reliably before in similar
situations (notably, Rez) and makes the experience of using tcsh/csh
better overall.

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>