Restore Config::csh_expect_file field
For interop between a new build of spk and an old build of spfs, bring this
field back so that when spk serializes this struct this field is present
(but with a dummy value).

spfs builds prior to #700 will fail to create a runtime if this field is
missing.

Mark the field with `serde(skip_deserializing)` so that while this field
is brought back, spfs will no longer _require_ the field be present. This
provides a migration path towards removing this field entirely after all
old installations of spfs have been updated.

At SPI our CI runs the test suite against whatever prod version of the
spfs+spk RPM is installed, rather than using a container like github CI
(for better or worse), so without these changes the test suite fails. We
also like to be able to invite testers to run newer versions of spk before
we dist out a new spfs+spk RPM.

Not all workflows are supported with this change. For example, if a runtime
is created with this change in a post-#700 spk, then using a pre-#700 spfs
build, it is not possible to `spfs join` the runtime:

    "/usr/bin/expect" "/dev/null" "/usr/bin/tcsh" "/tmp/spfs-runtime/startup.csh"

Signed-off-by: J Robert Ray <jrray@jrray.org>