Specialize diff calculation for package validation
Validating a new package build means checking if the package did anything
naughty like delete a file from a dependency. This validation process does
not need to build a manifest (hash the contents) of every file in /spfs,
which can take many minutes for large build environments. Instead, build
a manifest out of what is found in the overlayfs upperdir, which will limit
what gets hashed to what was really modified by the build.

This is accomplished by committing the upperdir just like what happens
after the validation phase. It would be a bonus to reuse some of this work
(like the content hash calculation) because as it stands right now it will
redo that work.

Signed-off-by: J Robert Ray <jrray@jrray.org>