Add tag namespace concept
From #390, the desire is to have different users (or CI jobs on the same
runner) have a certain amount of isolation from each other. In particular,
`spk ls` is expected to not contain any "local" packages that don't
belong to the current user/CI job.

This commit experiments with this idea. Only the local repo is expected to
have an optional tag namespace set, so tar and rpc repos have not been
changed. Setting the tag namespace is possible via the spfs config file,
but it is expected to typically be set via the environment variable
$SPFS_STORAGE_TAGNAMESPACE instead.

It may be reasonable to add a [spk] config option to make the tag
namespace default to the current user's username. A user sharing a host
with other users are unlikely to want to see any local packages made by
other users.

There is an open question on if spfs runtimes should be subject to the tag
namespace. This code currently does make them subject to the tag namespace.
Therefore, if a user has a tag namespace specified, `spfs runtime ls` would
only show the runtimes created in that same tag namespace, and those
runtimes would be hidden from other users/other tag namespaces. This is in
the same spirit of users being isolated from each other, however an
argument could be made that it is important from an admin perspective to
be able to list _all_ the runtimes on a host.

`spfs clean` is a special case that always ignores any tag namespace. It
should be able to find _all_ tags in the repo and avoid coming to any false
conclusions about what objects are garbage on the basis of only seeing
tags inside a tag namespace.

Signed-off-by: J Robert Ray <jrray@jrray.org>