Add more error context in render
Introduce a new `Error::StringWithSource` to wrap an error with a custom
context message that anyhow can make use of when displaying the error. But
we must be careful to not wrap errors in contexts where callers are
expecting certain error types. Having separate error types for specific
methods would help with this.

    jrray@STRANDED:~/git/spk$ spk env spk-convert-pip -- true
    ERROR render layers

    Caused by:
        0: render manifest N3PZ7AUENWBU7XOSOIH36KI4XLUY2KZ3LJQAVR3GEVU47WOO2LIQ====
        1: Unknown Object: HZQX6ZMXPVJF724OJEXTL2BUPLQKEGDBRYY5G7OYI4O33ZDF4UYA====
    ERROR Failed to spawn spfs-render

    Caused by:
        process exited with non-zero status
    ERROR Failed to re-mount runtime filesystem

Signed-off-by: J Robert Ray <jrray@jrray.org>