Also repair blobs in PayloadFallback
The question really is should we always run the syncer in a more expensive
mode where it verifies all objects really exist, or do we repair things as
they are discovered in a way that doesn't add any overhead?

A render can fail because of a missing blob object, but we're already using
a PayloadFallback repo and the missing blob is possible to be fetched from
the configured fallback(s). But the syncer has already skipped syncing the
blob because the manifest object already existed. It would be expensive to
verify every child object of a manifest really exists. But we will discover
which blobs are missing as things are rendered.

Now the name PayloadFallback is misleading because it does more than
repair payloads.

Signed-off-by: J Robert Ray <jrray@jrray.org>