Add a naming convention for tag namespace directories
Use a character that is an illegal character for users, '#', so we can
prevent users from creating tag names that could be misinterpreted as
namespace names.

To prevent different namespaces from getting mixed up, include the depth
of the namespace component in the directory name. For example the tag "hi"
in namespace "foo/bar" would produce this on disk:

    /spfs-repos/local/tags/foo#ns.0
    /spfs-repos/local/tags/foo#ns.0/bar#ns.1
    /spfs-repos/local/tags/foo#ns.0/bar#ns.1/hi.tag

This prevents tag "bar/hi" in namespace "foo" from colliding with tag "hi"
in namespace "foo/bar" since these would have different filenames:

- tags/foo#ns.0/bar/hi.tag
- tags/foo#ns.0/bar#ns.1/hi.tag

Signed-off-by: J Robert Ray <jrray@jrray.org>