Simple test of tag namespace behavior
The test code requires a way to modify the tag namespace of a TempRepo,
but since it can be changed through an Arc, add a `with_tag_namespace`
method to make a duplicate handle to an existing repo but with a different
namespace set. This required putting the TempDir inside an Arc.

Signed-off-by: J Robert Ray <jrray@jrray.org>