Replaces --solver-to-show with --solver-to-run (#696)
The '--solver-to-run' option specifies which solver to run and show
the output from. It defaults to 'all', which runs all the solvers
('cli' and 'checks') in parallel and shows the output from the 'cli'
solver. If given 'cli' or 'checks' it will run just that solver and
shows its output.

The "solver has finished" messages are updated to say whether the
solver that finished first found a solution or not. And, if a solver
whose output is hidden finished first, it will tell the user how to
rerun with just that solver.

Signed-off-by: David Gilligan-Cook <dcook@imageworks.com>