Fix spk test not looking for correct build digest
Revising #583, this was still not quite calculating a digest in the same
way that matches what BinaryPackageBuilder does. The missing piece was
calling `resolve_options`, and it happens that `self.options` already has
the variant's options baked into it, so we don't actually need to pass the
variant into PackageInstallTester.

Signed-off-by: J Robert Ray <jrray@imageworks.com>