Update monitor to tear down all mount with spfs-enter --exit
Overlayfs was holding on to the fuse filesystem in the lower dir and not
allowing the kernel to clean it up completely. The fix is to have the
spfs-enter command able to tear down mounts via the --exit flag, and
then have the monitor call this binary instead of only unmounting fuse.

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>