Refactor env to enforce permissions & namespace
Building off of the mount namespace guard, this introduces a generic
type that protects the privileged and namespace-dependent functions
behind processes which ensure that they can only be called once the
dependencies have been verified and/or the correct order of operations
has been followed.

This also means that fuse mounts are mounted as root and require root to
unmount them, but the new code enforces this appropriately. Primarily,
this removes the need for `user_allow_other` in the fuse config.

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>