Add marker traits for thread vs process
These marker traits allow for methods to be made available for both
`ThreadIsInMountNamespace` and `ProcessIsInMountNamespace` (for methods
that have callers from both flavors).

Return a `RuntimeConfigurator<_, ProcessIsInMountNamespace>` when possible
and for callers that always run in `ProcessIsInMountNamespace` flavor,
have those methods require this so the implementation does not need to
spawn a new thread. For example: `unmount_env`.

Signed-off-by: J Robert Ray <jrray@jrray.org>