Fix infinite loop attempting to repair missing payload
`FallbackProxy::open_payload` could get stuck in an infinite loop if a
payload is missing and the object with the same payload is not a blob
object (see #674 for more about hash collisions in the repo).

Don't return a `read_blob` error as `UnknownObject` if the object exists
but is not a blob. Preserve the `ObjectNotABlob` error. For added safety
to prevent looping forever, require the digest match an existing blob
before retrying `open_payload`.

Signed-off-by: J Robert Ray <jrray@jrray.org>