Handle package with overlapping requirements
This can be seen with opencv-contrib-python-headless==4.7.0.72, it has
multiple `Requires-Dist` entries for different versions of numpy. This was
getting converted into multiple entries in spk's install.requirements and
then the spec was rejected as illegal:

    ERROR
     1 | ...}, "install": {"requirements": [{"pkg": "python/3.9"}, {"pkg"...
       |                                   ^ install.requirements: found multiple install requirements for 'python-numpy' at line 1 column 419

Avoid this by combining multiple requirements into a single one:

    install:
      requirements:
        - pkg: python/3.9
        - pkg: "python-numpy/>=1.17.0,>=1.17.3,>=1.19.3"

This means the code here doesn't have to be clever about combining these,
and the spk code will simplify this request into just `>=1.19.3`.

Signed-off-by: J Robert Ray <jrray@imageworks.com>