Replace pin bool with enum
The VarRequest had a pin field used to denote when fromBuildEnv was
given and this field was mutually exclusive to a provided value. An enum
works better in this case. Ideally, the package would not allow this
field to be pinned, only the recipe. For now, I have not made that
change only this first one to introduce an enum.

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>