Fix `spk build` build digest calculation
Like in #729 and #730, `spk test` may fail to solve because it is using a
different code path to predict the appropriate build digest for the variant
being tested. In this case, `spk build` was using resolved options while
calculating the build digest, but `spk test` does not. Update `spk build`
to also calculate the build digest based on the non-resolved options, as
intended. `spk build` should always produce the same build digest for the
same input recipe, on the same host, regardless of what the solver finds
in the repo.

Signed-off-by: J Robert Ray <jrray@imageworks.com>