Prevent building packages that have circular dependencies
As demonstrated in the new test, it was previously possible to build a
package "one" that depends on a package "two" that depends back on package
"one" -- provided that a build of "one" already exists in the repo.

An odd side effect of this being possible before is that during the build
the options would include things like `one.two: ~1.0.0` and would influence
the build hash, because instead of the variant's options including `two: `
it would include `two: ~1.0.0`. Then, `spk test` would calculate a
different build hash for the same variant, because it would use `two: `.

Signed-off-by: J Robert Ray <jrray@jrray.org>