Use Arc<str> instead of String for OptionMap value
First in a series of commits to make this kind of change for values that
are part of `State` and will benefit greatly from faster cloning and lower
memory footprint. Cloning an Arc is an O(1) operation and the solver has
to make many clones of State as it progresses.

There is currently a lot of `.to_string()` used on these values but this
is expected to change into Arc::clone as other things get similarly
converted.

Signed-off-by: J Robert Ray <jrray@jrray.org>