Use string interning for PkgNameBuf, etc.
Use the lasso crate to intern package name strings, so that any instances
of the same name share the same backing memory. These are also very cheap
to clone because it only clones the "Spur" (a u32).

The tradeoff is that there is some contention from using `ThreadedRodeo`
since we need to support accessing the backing store from multiple threads.

Signed-off-by: J Robert Ray <jrray@jrray.org>