Support python package "extras"
A number of issues came up while trying to spk-covert-pip "gradio" for
python 3.9.

It started out with there being a bug where the spk package install
requirements were getting generated incorrectly, including an entry like:

    - pkg: python-markdown-it-py[linkify]/3.0.0

The "linkify" here is a package extra and it wasn't being stripped from
the package name.

When there is a pip request, or dependency requirement, that includes
extras, as in:

    markdown-it-py[linkify]==3.0.0

support this by _not_ filtering out the requirements that match this extra
value, and generate an spk package var with this extra declared as enabled,
as in:

    - var: python_extra_linkify/true

Then, this package might have extra install.requirements that wouldn't be
present if that extra wasn't enabled. If someone spk-convert-pip's this
package again without the extra enabled, it would generate a different
build without this var and with a different build hash. This way those
packages that depend on this package with this extra enabled will pick
the build with this var set and bring in the extra dependencies.

The packages generated that have a dependency on a package with extras
enabled will get the appropriate var(s) defined as install requirements. As
seen in the python-gradio package:

    - var: python-markdown-it-py.python_extra_linkify/true

The `gradio` package has a requirement for `markdown-it-py[linkify]`.

Then, a really strange situation presented itself. gradio has these two
requirements:

    [1] markdown-it-py[linkify]>=2.0.0 (resolves to 3.0.0 w/ linkify)

    [2] mdit-py-plugins<=0.3.3 -> markdown-it-py>=1.0.0,<3.0.0 (resolves to 2.2.0 w/o linkify)

When processing [1], it will create markdown-it-py/3.0.0 with the linkify
extra enabled, but [2] requires markdown-it-py < 3.0.0. Two things need to
happen here. When visiting mdit-py-plugins again, it needs to not be
skipped, because the requirement has changed from ">=2.0.0" to
">=1.0.0,<3.0.0" and resolves to a different version. Then, when packaging
mdit-py-plugins 2.2.0 for requirement [2], the linkify extra needs to be
enabled so the produced build can still satisfy the requirement from [1]
for `markdown-it-py[linkify]`.

Instead of trying to detect when a repeat request for the same package name
would resolve to the same downloaded version, this takes the easier
appoach of running `pip download` again with the new request and seeing if
it downloads the same file. This avoids having to maintain extra logic at
the cost of doing redundant downloads.

Signed-off-by: J Robert Ray <jrray@imageworks.com>